Move lms.lib.xblock into lms.djangoapps.lms_xblock in preparation add database backed configuration to it
