+++ lms/djangoapps/courseware/tests/test_lti_integration.py
@@ -13,7 +13,7 @@ from django.test.utils import override_settings
 from courseware.tests import BaseTestXmodule
 from xmodule.modulestore.tests.django_utils import TEST_DATA_MOCK_MODULESTORE
 from courseware.views import get_course_lti_endpoints
-from lms.lib.xblock.runtime import quote_slashes
+from lms.djangoapps.lms_xblock.runtime import quote_slashes
 from xmodule.modulestore.tests.django_utils import ModuleStoreTestCase
 from xmodule.modulestore.tests.factories import CourseFactory, ItemFactory
 from xmodule.x_module import STUDENT_VIEW
